Proof of Concept: Separate scopes out into own objects
