+++ app/policies/application_policy.rb
@@ -1,12 +1,4 @@
 class ApplicationPolicy
-  class Scope
-    attr_reader :user, :scope
-
-    def initialize(user, scope)
-      @user = user
-      @scope = scope
-    end
-  end
 
   attr_reader :user, :record
