+++ app/policy_scopes/application_policy_scope.rb
@@ -1,0 +1,10 @@
+class ApplicationPolicyScope
+
+  attr_reader :user, :scope
+
+  def initialize(user, scope)
+    @user = user
+    @scope = scope
+  end
+
+end
