+++ app/policy_scopes/user_policy_scope.rb
@@ -1,0 +1,11 @@
+class UserPolicyScope < ApplicationPolicyScope
+
+  def resolve
+    if user.present? && (user.admin? || user.member?)
+      scope
+    else
+      scope.none
+    end
+  end
+
+end
