+++ spec/policy_scopes/application_policy_scope_spec.rb
@@ -1,0 +1,19 @@
+require 'rails_helper'
+
+describe ApplicationPolicyScope do
+  let(:policy_scope) { ApplicationPolicyScope.new(user, scope) }
+  let(:user)  { User.new }
+  let(:scope) { User.all }
+
+  describe "#initialize" do
+    subject { policy_scope }
+
+    it "sets the user" do
+      expect(subject.user).to eq(user)
+    end
+
+    it "sets the scope" do
+      expect(subject.scope).to eq(scope)
+    end
+  end
+end
